+++ src/tests/Equinor.Procosys.Preservation.Command.Tests/TagCommands/CreateAreaTag/CreateAreaTagCommandTests.cs
@@ -38,150 +38,6 @@ namespace Equinor.Procosys.Preservation.Command.Tests.TagCommands.CreateAreaTag
         }
 
         [TestMethod]
-        public void GetTagNo_ShouldReturnPreTagNo_WithDiscipline()
-        {
-            var dut = new CreateAreaTagCommand(
-                "",
-                TagType.PreArea,
-                "I",
-                null,
-                null,
-                0,
-                null,
-                null,
-                null,
-                null);
-
-            Assert.AreEqual("#PRE-I", dut.GetTagNo());
-        }
-
-        [TestMethod]
-        public void GetTagNo_ShouldReturnSiteTagNo_WithDisciplineAndSuffix()
-        {
-            var dut = new CreateAreaTagCommand(
-                "",
-                TagType.SiteArea,
-                "I",
-                null,
-                null,
-                0,
-                null,
-                null,
-                null,
-                null);
-
-            Assert.AreEqual("#SITE-I", dut.GetTagNo());
-        }
-
-        [TestMethod]
-        public void GetTagNo_ShouldReturnPreTagNo_WithDisciplineAndSuffix()
-        {
-            var dut = new CreateAreaTagCommand(
-                "",
-                TagType.PreArea,
-                "I",
-                null,
-                "XX",
-                0,
-                null,
-                null,
-                null,
-                null);
-
-            Assert.AreEqual("#PRE-I-XX", dut.GetTagNo());
-        }
-
-        [TestMethod]
-        public void GetTagNo_ShouldReturnSiteTagNo_WithDiscipline()
-        {
-            var dut = new CreateAreaTagCommand(
-                "",
-                TagType.SiteArea,
-                "I",
-                null,
-                "XX",
-                0,
-                null,
-                null,
-                null,
-                null);
-
-            Assert.AreEqual("#SITE-I-XX", dut.GetTagNo());
-        }
-
-        [TestMethod]
-        public void GetTagNo_ShouldReturnPreTagNo_WithDisciplineAndArea()
-        {
-            var dut = new CreateAreaTagCommand(
-                "",
-                TagType.PreArea,
-                "I",
-                "A300",
-                null,
-                0,
-                null,
-                null,
-                null,
-                null);
-
-            Assert.AreEqual("#PRE-I-A300", dut.GetTagNo());
-        }
-
-        [TestMethod]
-        public void GetTagNo_ShouldReturnSiteTagNo_WithDisciplineAndArea()
-        {
-            var dut = new CreateAreaTagCommand(
-                "",
-                TagType.SiteArea,
-                "I",
-                "A300",
-                null,
-                0,
-                null,
-                null,
-                null,
-                null);
-
-            Assert.AreEqual("#SITE-I-A300", dut.GetTagNo());
-        }
-
-        [TestMethod]
-        public void GetTagNo_ShouldReturnPreTagNo_WithDisciplineAndAreaAndSuffix()
-        {
-            var dut = new CreateAreaTagCommand(
-                "",
-                TagType.PreArea,
-                "I",
-                "A300",
-                "XX",
-                0,
-                null,
-                null,
-                null,
-                null);
-
-            Assert.AreEqual("#PRE-I-A300-XX", dut.GetTagNo());
-        }
-
-        [TestMethod]
-        public void GetTagNo_ShouldReturnSiteTagNo_WithDisciplineAndAreaAndSuffix()
-        {
-            var dut = new CreateAreaTagCommand(
-                "",
-                TagType.SiteArea,
-                "I",
-                "A300",
-                "XX",
-                0,
-                null,
-                null,
-                null,
-                null);
-
-            Assert.AreEqual("#SITE-I-A300-XX", dut.GetTagNo());
-        }
-
-        [TestMethod]
         public void Constructor_ShouldSetProperties_WithNullLists()
         {
             var dut = new CreateAreaTagCommand("ProjectNameA",
